integrated cpu estimation into monitor

+++ energy_monitor/powergadget_wrapper.py
@@ -23,16 +23,18 @@ class monitor:
     def __init__(self, name='Monitor Test',
                        non_standard_binary_path=False,
                        remove_background_energy=False,
-                       log_filepath=os.path.join(os.path.expanduser("~"), 'Documents', 'energy_monitor_log.csv')):
+                       log_filepath=os.path.join(os.path.expanduser("~"), 'Documents', 'energy_monitor_log.csv'),
+                       TDP=15):
         self.system_os = platform.system()
         self.name = name
         self.non_standard_binary_path = non_standard_binary_path
         self.remove_background_energy = remove_background_energy
         self.log_filepath = log_filepath
+        self.TDP = TDP
         self.app_running = False
         self.recording = False
         self._locate_bin()
-        self.power_gadget_data = {}
+        self.write_data = {}
         # TODO: check log filepath is '.csv' format
 
     def _locate_bin(self):
@@ -41,18 +43,24 @@ class monitor:
                 raise ValueError('argument: non_standard_binary_path={arg} needs to be type str'.format(arg=repr(self.non_standard_binary_path)))
             else:
                 self.bin = self.non_standard_binary_path
-            # TODO: now test the bin file exists
+                # TODO: now test the bin file exists
+                if os.path.exists(self.bin):
+                    self.IntelPowerGadget = True
+                else:
+                    self.IntelPowerGadget = False
         else: # use standard application binary location
             if self.system_os == 'Windows':
                 self.bin = '"C:\\Program Files\\Intel\\Power Gadget 3.6\\IntelPowerGadget.exe"'
-            elif self.system_os == 'Darwin':
-                self.bin = '/Applications/Intel/IntelPowerGadget.bin'
+                self.IntelPowerGadget = True
             else:
-                raise OSError('Intel Power Gadget not support on operating system: {arg}'.format(arg=repr(self.system_os)))
+                self.IntelPowerGadget = False
 
     def _start_cmd(self):
-        self._open_app() # make sure IntelPowerGadget is running
-        os.system(self.bin + ' -start')
+        if self.IntelPowerGadget:
+            self._open_app() # make sure IntelPowerGadget is running
+            os.system(self.bin + ' -start')
+        else: # use estimation
+            self.cpu_monitor = cpu_percent.start_recording(TDP=self.TDP)
         self.start_time = datetime.now() # log start time to know file name
         self.recording = True
 
@@ -83,19 +91,22 @@ class monitor:
 
     def _stop_read_delete(self):
         if self.recording:
-            os.system(self.bin + ' -stop')  # CLI stop recording
-            while not self._check_PwrData_csv_exists(): # wait for csv to be created
-                time.sleep(0.1)
-            while not utils.check_written(self.csv_file): # wait for csv to be written
-                time.sleep(0.1)
+            if self.IntelPowerGadget:
+                os.system(self.bin + ' -stop')  # CLI stop recording
+                while not self._check_PwrData_csv_exists(): # wait for csv to be created
+                    time.sleep(0.1)
+                while not utils.check_written(self.csv_file): # wait for csv to be written
+                    time.sleep(0.1)
+                # save what we need from csv
+                self.data = utils.read_csv(self.csv_file)
+                # close recording
+                self._kill_proc()
+                # remove csv log
+                os.remove(self.csv_file)
+            else:
+                self.data = self.cpu_monitor.stop()
+            self.joules = self.data['cumulative_ia']
             self.recording = False
-        # save what we need from csv
-        self.power_gadget_data = utils.read_csv(self.csv_file)
-        self.joules = self.power_gadget_data['cumulative_ia']
-        # close recording
-        self._kill_proc()
-        # remove csv log
-        os.remove(self.csv_file)
 
     def stop(self):
         '''
@@ -104,14 +115,17 @@ class monitor:
         '''
         self._stop_read_delete()
         # log monitor data
-        self.power_gadget_data['date'] = utils.get_date_string(self.start_time)
-        self.power_gadget_data['name'] = self.name
-        self.power_gadget_data['remove_background_energy'] = str(self.remove_background_energy)
+        self.write_data = {}
+        self.write_data['cumulative_ia'] = self.data['cumulative_ia']
+        self.write_data['average_ia'] = self.data['average_ia']
+        self.write_data['date'] = utils.get_date_string(self.start_time)
+        self.write_data['name'] = self.name
+        self.write_data['remove_background_energy'] = str(self.remove_background_energy)
         # remove background
-        self.power_gadget_data['cumulative_ia'] -= self.background_watts*self.power_gadget_data['time']
-        self.power_gadget_data['average_ia'] -= self.background_watts
+        self.write_data['cumulative_ia'] -= self.background_watts*self.data['time']
+        self.write_data['average_ia'] -= self.background_watts
         # log
-        utils.log_data(self.log_filepath, self.power_gadget_data)
+        utils.log_data(self.log_filepath, self.write_data)
 
     def _check_PwrData_csv_exists(self):
         for file in utils.get_PwrData_csv(self.start_time):
@@ -126,7 +140,7 @@ class monitor:
             self.app_running = False
 
     def __str__(self):
-        return str(self.power_gadget_data)
+        return str(self.write_data)
 
     def __enter__(self):
         # for use with 'with' statement enter
